requirements.txt
gradio>=4.0.0
pillow
requests
fal-client
python-dotenv
numpy
scikit-image
